server/requirements1.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albumentations==1.1.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cycler==0.12.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.4.3
mediapipe==0.10.0
multidict==6.1.0
networkx==3.1
numpy==1.18.5
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pgi==0.0.11.2
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyDeprecate==0.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.9
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.9.3
six==1.16.0
sounddevice==0.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==1.9.0
torchmetrics==0.11.4
torchvision==0.10.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.1
